+++ tests/php/Unit/Api/Application/PhpSymbolCatalogTest.php
@@ -7,6 +7,8 @@ namespace PhelTest\Unit\Api\Application;
 use Phel\Api\Application\PhpSymbolCatalog;
 use PHPUnit\Framework\TestCase;
 
+use function array_keys;
+
 final class PhpSymbolCatalogTest extends TestCase
 {
     public function test_functions_contains_internal_php_functions(): void
@@ -40,5 +42,27 @@ final class PhpSymbolCatalogTest extends TestCase
         $catalog = new PhpSymbolCatalog();
 
         self::assertSame($catalog->classes(), $catalog->classes());
+    }
+
+    public function test_superglobals_cover_the_language_defined_set(): void
+    {
+        self::assertSame([
+            '$GLOBALS',
+            '$_SERVER',
+            '$_GET',
+            '$_POST',
+            '$_FILES',
+            '$_COOKIE',
+            '$_SESSION',
+            '$_REQUEST',
+            '$_ENV',
+        ], array_keys(PhpSymbolCatalog::SUPERGLOBALS));
+    }
+
+    public function test_every_superglobal_carries_a_description(): void
+    {
+        foreach (PhpSymbolCatalog::SUPERGLOBALS as $variable => $description) {
+            self::assertNotSame('', $description, $variable . ' has no description');
+        }
     }
 }